requirements.txt
pyrogram
tgcrypto
pytz
pytgcalls==3.0.0.dev24
yt-dlp
youtube-search-python
motor==3.0.0
dnspython
GitPython
pykeyboard
bs4
Pillow
pymediainfo
requests
unidecode
attrify
pyromod==1.1.0
html2text
removebg
aiofiles
aiohttp
gpytranslate
gtts
wget
telegraph
heroku3
openai
APScheduler
kynaylibs==1.5.8
psutil